fix: 파일 중복 및 변경된 DTO 수정 #62
